refactor: extract shared AudioVideoSettingsPanel component
Deduplicate audio/video settings between VoiceSettings page and
DeviceSettingsDialog by extracting a shared AudioVideoSettingsPanel.
The dialog now has full settings parity (input mode, PTT key, voice
activity threshold) instead of only device selectors.
